# backend/app/api/swagger/spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Timeline API",
        "version": "1.0.0",
        "description": "Comprehensive REST API for Timeline project management application. Manage workspaces, projects, tasks, team members, and collaboration features.",
        "contact": {
            "name": "Timeline API Support",
            "email": "api-support@timeline.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://timeline.iamrakesh.codes/terms"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "x-meta": {
                "pathsCount": 46,
                "endpointOperations": 54,
                "models": 8,
                "services": 14,
                "countingRules": "Excluded: /swagger, /swagger-ui, / (root redirect). Included AI endpoints and OAuth endpoints. Counted distinct path templates as pathsCount; each HTTP method on a path counted in endpointOperations."
            },
            "description": "Development server"
        },
        {
            "url": "https://not-known.com/v1",
            "description": "Production server"
        },
        {
            "url": "https://staging-api.timeline.com/v1",
            "description": "Staging server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token authentication. Format: 'Bearer {token}'"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "__session",
                "description": "Session-based authentication using HTTP-only cookies"
            },
            "clerkAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Clerk.js authentication token"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API key for server-to-server authentication"
            }
        },
...
    },
    "paths": {
...
        "/users/stats": {
            "get": {
                "summary": "User statistics",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "User statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "totalUsers": {
                                            "type": "integer"
                                        },
                                        "activeUsers": {
                                            "type": "integer"
                                        },
                                        "newUsersToday": {
                                            "type": "integer"
                                        },
                                        "premiumUsers": {
                                            "type": "integer"
                                        },
                                        "usersByRole": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        },
                                        "recentActivity": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "lastLoginAt": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updatedAt": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "role": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "availableFields": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "lastUpdated": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/users/count": {
            "get": {
                "summary": "Basic user counts",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Counts fetched"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/database/stats": {
            "get": {
                "summary": "Database statistics",
                "tags": [
                    "Database"
                ],
                "responses": {
                    "200": {
                        "description": "Stats returned"
                    },
                    "503": {
                        "description": "Database unavailable"
                    }
                }
            }
        },
...
    }
}